+++ es.sidelab.pascaline.cdtinterface/src/es/sidelab/pascaline/cdtinterface/ui/wizards/FreePascalMainWizardPage.java
@@ -45,6 +45,9 @@ public class FreePascalMainWizardPage extends CDTMainWizardPage {
 					filteredItems.add(ed);
 				}									
 			}
+			if(ed.getId().equals("Others")) {
+				filteredItems.add(ed);
+			}			
 		}
 		return filteredItems; 
 	}
